stocks_websocket_server/requirements.txt
aiohttp==3.7.4
alpaca-trade-api==1.4.3
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
Flask==2.0.2
Flask-Executor==0.10.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
msgpack==1.0.2
multidict==5.2.0
numpy==1.22.0
pandas==1.3.5
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.8
websocket-client==1.2.3
websockets==9.1
Werkzeug==2.0.2
yarl==1.7.2
